revert(view): move files
